Add content and basic structure
The bulk of the HTML is built. Adding layout might result in some changes to
the HTML. Some of the style information is also added, as it came to mind,
though that was not the objective at this point.

Except for the listing information, the textual content is place, with three of
the listings, choosen for their content, are placed. Having three, the max row
size, included will allow me to develop the layout controlling CSS.

One decision made at this point is using in-element style tags for the listing
photos. I want to place the photos as background images. Firstly as the style
for background images is easier to control for layout, and secondly it is a bit
harder for users to "copy" background images. The trade off is that they are not
printable either. I will work on that issue, if I have time. This is, seemingly,
counter to the intended split between HTML for content and CSS for style. As a
point, it really is not counter to that as the listing photos are, in fact, part
of the content, and likely to be database-driven for a live site. My personal
bias is against having a stylesheet, internal or external, being built in a
dynamic fashion based on content. Placing the photo URLs in the styles would
set up a situation where that would have to happen, for every page generated.
The HTML generation code, whatever it might be, can add the URL for the photo
to the in-element style declaration, allowing all other aspects to be under
the control of a static CSS, or even a SASS generated version of the CSS.

Have to wait and see if the instructors, a) notice, and b) ding me for this.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>