Start developing the layout
Changed some of the elements for semantic purposes, and added more depth to some
of the sections. Mapping out aria attributes on menu items.

There are style sections scattered through the body so that editing the styles
can be done closer to the content being styled. Part of the final processing is
to gather these and place them in the html head section.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>