Add more currency symbols
Changed the way the currency symbol is selected, with unique class names. Added
another font, to handle BTC. Now, as long as the symbol preceeds the price, any
currency can be added using a simple class, and possibly linking to a font-face
which has that symbol included.

Doubled up the sample listing count, for upcoming grid work, and to show how to
add a current symbol to the page.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>