Add the supplied data
Replace the four sample listings with the data supplied by the starter code.

Remove the sample data from the comments.

Corrected some of the supplied data to match with the solution given as a guide,
and alter some of the other values to activate some of the features built into
this version of the page.

Not that they all make sense, but added the accesskey properties as well.

Add some responsive work to the main menu, more would be needed for a live site.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>