Rework theme control system
The objective is to remove the color controls from the individual declarations
and place it in a single place and using variables to switch the colors.

Using an individual theme test combined with each declaration was cumbersome
to build and would be unrealistic to maintain.

Using CSS variables allows me to create all the color choices in a single place
and use the results throught the declartions. As a bonus, by using meaningful
names, all the colors used within the declarations indicate the purpose of the
color rather than merely what color it is.

Alongside the process of switching the color use around I was also able to drop
a significant amount of the spurious class and id usage.

Still to do, if ever done, is to select more colors for the dark theme. As a
demonstration of the possibilty, it works. As a "theme" is falls flat.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>